+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.0 RC1 1194 (2018-11-01)
 
 ### Added
 
@@ -47,7 +47,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reject unrecognized values for `cipher`, `auth` and `proto`. [#1](https://github.com/keeshux/passepartout-ios/pull/1)
 - Alert unsupported configuration options.
 - Use accent color for checkmarks in table cells.
-- Hosts gone while connected. [#19](https://github.com/keeshux/passepartout-ios/issues/19)
 
 ## 1.0 beta 975 (2018-10-11)